Запомни эту фразу, всё будет, но не сразу.
